Rename directory and update all imports
