+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:simple_chat/configuration/config_invoke.dart';
 import 'package:simple_chat/configuration/countries_invoke.dart';
 import 'package:simple_chat/utils/logger_config.dart';
 import 'package:simple_chat/screens_pages/log_in_page.dart';
-import 'package:simple_chat/methods_functions/user_entrypoint_methods.dart';
+import 'package:simple_chat/standalone_methods_functions/user_entrypoint_methods.dart';
 
 //imports
 /////////////////////////////////////////////////////////////////////////////
