+++ lib/screens_pages/complete_profile.dart
@@ -4,8 +4,8 @@ import 'package:google_fonts/google_fonts.dart';   //Fonts
 import 'package:intl/intl.dart';
 import 'package:simple_chat/configuration/countries_invoke.dart';
 
-import 'package:simple_chat/methods_functions/general_methods.dart';
-import 'package:simple_chat/methods_functions/user_entrypoint_methods.dart';
+import 'package:simple_chat/standalone_methods_functions/general_methods.dart';
+import 'package:simple_chat/standalone_methods_functions/user_entrypoint_methods.dart';
 import 'package:simple_chat/configuration/config_invoke.dart';
 import 'package:simple_chat/screens_pages/landing_page.dart';
 import 'package:simple_chat/utils/alert_dialog_list.dart';
