+++ lib/screens_pages/sign_up_page.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';   //Fonts
 
-import 'package:simple_chat/methods_functions/general_methods.dart';
-import 'package:simple_chat/methods_functions/user_entrypoint_methods.dart';
+import 'package:simple_chat/standalone_methods_functions/general_methods.dart';
+import 'package:simple_chat/standalone_methods_functions/user_entrypoint_methods.dart';
 import 'package:simple_chat/configuration/config_invoke.dart';
 import 'package:simple_chat/utils/alert_dialog_list.dart';
 import 'package:simple_chat/utils/logger_config.dart';
