+++ lib/standalone_methods_functions/user_entrypoint_methods.dart
@@ -6,7 +6,7 @@ import "dart:async";
 import "dart:convert";
 import 'package:http/http.dart' as http;
 import "package:simple_chat/session_related/app_storage_class.dart";
-import "package:simple_chat/methods_functions/general_methods.dart";
+import "package:simple_chat/standalone_methods_functions/general_methods.dart";
 
 //Import alert dialogs and others
 import "package:simple_chat/utils/alert_dialog_list.dart";
